# storage/api-docs/api-docs.yaml
...
openapi: "3.0.0"
info:
  title: "Upstart API"
  version: "1.0.0"
paths:
...
  /api/v1/categories:
    get:
      tags:
        - "Category"
      summary: "Get Categories"
      operationId: "Category"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: en
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  properties:
                    id:
                      type: "integer"
                    parent_id:
                      type: "integer"
                    ordering:
                      type: "integer"
                    title:
                      type: "string"
                    icon:
                      type: "string"
                    child:
                      type: "array"
                      items:
                        properties:
                          id:
                            type: "integer"
                          parent_id:
                            type: "integer"
                          ordering:
                            type: "integer"
                          title:
                            type: "string"
                          icon:
                            type: "string"
                        type: "object"
                  type: "object"
        "401":
          description: "Unauthenticated"
...
  /api/v1/moderators:
    get:
      tags:
        - "Moderator"
      summary: "Get Moderators"
      operationId: "Moderator"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: hy
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/Moderator'
        "401":
          description: "Unauthenticated"
      security:
        - bearerAuth: [ ]
...
  /api/v1/moderators/create:
    post:
      tags:
        - "Moderator"
      operationId: "moderator"
      description: This request is only for super Admin users. In  the request headers  need set user api_token as the bearer token.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Moderator'
          application/xml:
            schema:
              $ref: '#/components/schemas/Moderator'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Moderator'
        required: true

      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: "integer"
                  first_name:
                    type: "string"
                  last_name:
                    type: "string"
                  email:
                    type: "string"
                  role_id:
                    type: "integer"
                  company_name:
                    type: "string"
                  tax_identity_number:
                    type: "integer"
                type: "object"
        "401":
          description: "Unauthenticated"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    Moderator:
      type: object
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        company_name:
          type: "string"
        role_id:
          type: integer
          default: 2
        tax_identity_number:
          type: "integer"
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT